add the results from the code
Note1: For some reason my computer did not read the CSV format that is in the folder. I had to open using the excel format.
Note2: I have to adjust the final txt results, the header from my last dataset was different so I had to match them.
Note3: The code take around 10 min to run all the 60 individuals curve in the loop.
Note4: The unique ID is not separating the traditional from DAT, I create a new one called "unique_id"